Greatly simplified the algorithms for moving the shapes.

+++ include/c_shapes.h
@@ -7,7 +7,7 @@
 // 8 = bottom
 
 #define numShapes 7
-unsigned int c_shapes[numShapes][8] {
+unsigned short c_shapes[numShapes][8] {
 	{00, 00, 00, 00,
 	 13, 12, 12, 14},
